Implemented tax service. closes #5. Writing test for tax service #6

+++ src/main/java/com/interview/salestaxes/api/model/ItemReceiptDTO.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @Setter
 @Getter
@@ -14,7 +13,7 @@ public class ItemReceiptDTO {
 
     private Integer amount;
     private ProductDTO product;
-    private Double tax;
-    private BigDecimal priceTaxed;
+    private double tax;
+    private double priceTaxed;
 
 }
